// frontend/types/api-types.ts
export interface components {
    schemas: {
        BaseCompetitionDto: {
            /** @description Name of the competition */
            name: string;
            /** @description Abbreviation */
            abbr?: string;
            /** @description Sport ID */
            sportId?: number;
            /** @description Note */
            note?: string;
            /** @description Status of the competition */
            status: string;
            /**
             * Format: date-time
             * @description Start date
             */
            startDate?: string;
            /**
             * Format: date-time
             * @description End date
             */
            endDate?: string;
            /** @description Order */
            ord?: number;
            /** @description Base64 of the competition logo */
            base64?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}